Move time functions to own package
