+++ server/timeutils/timeutils_test.go
@@ -1,4 +1,4 @@
-package command
+package timeutils
 
 import (
 	"testing"
@@ -6,42 +6,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 )
-
-func TestGetTimeForMillis(t *testing.T) {
-	t.Run("Now", func(t *testing.T) {
-		expected := time.Now()
-
-		millis := expected.Unix() * 1000
-		actual := getTimeForMillis(millis)
-
-		require.Equal(t, expected.Unix(), actual.Unix())
-	})
-
-	t.Run("Epoch 0", func(t *testing.T) {
-		expected, err := time.Parse("2006-01-02 15:04:05 MST", "1970-01-01 00:00:00 UTC")
-		require.NoError(t, err)
-
-		actual := getTimeForMillis(0)
-
-		require.Equal(t, expected.Unix(), actual.Unix())
-	})
-
-	t.Run("Go layout", func(t *testing.T) {
-		// From https://golang.org/pkg/time/#pkg-constants:
-		// "The reference time used in the layouts is the specific time:
-		// Mon Jan 2 15:04:05 MST 2006
-		// which is Unix time 1136239445"
-		var unixTime int64 = 1136239445
-		layout := "Mon Jan 02 15:04:05 -0700 2006"
-
-		expected, err := time.Parse(layout, layout)
-		require.NoError(t, err)
-
-		actual := getTimeForMillis(unixTime * 1000)
-
-		require.Equal(t, expected.Unix(), actual.Unix())
-	})
-}
 
 func TestDurationString(t *testing.T) {
 	now := time.Now()
@@ -158,7 +122,7 @@ func TestDurationString(t *testing.T) {
 
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			actual := durationString(testCase.start, testCase.end)
+			actual := DurationString(testCase.start, testCase.end)
 			require.Equal(t, testCase.expected, actual)
 		})
 	}
